fix opaque types and extrinsics format
